+++ CHANGELOG.md
@@ -53,6 +53,9 @@ List of major changes and improvements
 
   - Change `LoopBlockingScheme` interface for blocking factors and loop orders.
 
+  - Loop blocking exhaustive search introduces regularized schemes and
+    suboptimal schemes, to enable more skips.
+
 - Software engineering.
 
   - Lazily evaluate loop blocking stats.